refactor: create common EmptyState component
